successfully made userForm a separate component
